Move text-feature warning into detect_feature_modalities
Per review feedback, the likely-text warning now fires from inside
detect_feature_modalities instead of being called separately by each
estimator. This gives a single call site (the one place TEXT labels are
produced) and a single place to test, so the classifier/regressor tests
collapse into one parametrized test under the modality-detection suite.

warn_if_text_features moves from validation to modality_detection;
stacklevel goes 5 -> 6 for the extra detect_feature_modalities frame.
