+++ changelog/1159.added.md
@@ -1,1 +1,1 @@
-`fit()` now warns when a column of `X` looks like free text: a string column with more than 30 distinct values that cannot be parsed as numbers. TabPFN has no text understanding, so such columns were silently encoded as near-unique integer codes, which adds noise and hurts performance. The warning names the affected columns and points to the fixes: convert the column to a numeric dtype if it holds numbers stored as strings, use the [tabpfn-client](https://github.com/PriorLabs/tabpfn-client) API if it holds genuine text, or pass the column's index in `categorical_features_indices` if it really is a high-cardinality category (which also silences the warning).
+`fit()` now warns when a column of `X` looks like free text.
