+++ src/tabpfn/preprocessing/modality_detection.py
@@ -4,6 +4,7 @@
 
 from __future__ import annotations
 
+import warnings
 from collections.abc import Sequence
 from typing import TYPE_CHECKING
 
@@ -11,6 +12,7 @@ import pandas as pd
 
 from tabpfn.errors import TabPFNUserError
 from tabpfn.preprocessing.datamodel import (
+    INPUT_FEATURE_PREFIX,
     Feature,
     FeatureModality,
     FeatureSchema,
@@ -21,6 +23,10 @@ if TYPE_CHECKING:
     import numpy as np
 
 _EARLY_EXIT_PREFIX_ROWS = 1024
+
+#: Cap on how many column names the likely-text warning lists, so a wide frame of
+#: text columns does not produce an unreadable multi-kilobyte message.
+_MAX_TEXT_COLUMNS_IN_WARNING = 10
 
 
 def detect_feature_modalities(
@@ -75,7 +81,73 @@ def detect_feature_modalities(
             big_enough_n_to_infer_cat=big_enough_n_to_infer_cat,
         )
         features.append(Feature(name=feature_name, modality=feat_modality))
-    return FeatureSchema(features=features)
+    feature_schema = FeatureSchema(features=features)
+    # Warn here rather than at each call site: this is the single place the TEXT
+    # labels are produced, and they do not survive the first preprocessing step
+    # that rebuilds the schema.
+    warn_if_text_features(
+        feature_schema,
+        declared_categorical_indices=provided_categorical_indices,
+    )
+    return feature_schema
+
+
+def warn_if_text_features(
+    feature_schema: FeatureSchema,
+    *,
+    declared_categorical_indices: Sequence[int] | None = None,
+) -> None:
+    """Warn when input columns look like free text rather than categoricals.
+
+    High-cardinality string columns are labelled `FeatureModality.TEXT` by
+    `detect_feature_modalities`, but this package has no text handling: they are swept
+    into the same `OrdinalEncoder` as real categoricals, which selects columns by dtype
+    (see `get_ordinal_encoder`). That turns near-unique text into near-unique integer
+    codes, i.e. noise rather than signal, without any error to hint at it.
+
+    Called by `detect_feature_modalities` while the schema still carries the TEXT
+    labels, i.e. before the first preprocessing step that rebuilds it, since
+    `FeatureSchema.from_only_categorical_indices` collapses TEXT into NUMERICAL.
+
+    Args:
+        feature_schema: The schema produced by `detect_feature_modalities`.
+        declared_categorical_indices: Positional indices the caller passed as
+            `categorical_features_indices`. These are never reported: declaring a
+            column categorical states that the user already knows it holds
+            non-numeric values and intends them as categories, so warning about it
+            would be noise.
+    """
+    declared = set(declared_categorical_indices or ())
+    text_names = [
+        feature.name.removeprefix(INPUT_FEATURE_PREFIX)
+        for index, feature in enumerate(feature_schema.features)
+        if feature.modality is FeatureModality.TEXT and index not in declared
+    ]
+    if not text_names:
+        return
+
+    shown = text_names[:_MAX_TEXT_COLUMNS_IN_WARNING]
+    column_names_to_print = ", ".join(repr(name) for name in shown)
+    if len(text_names) > len(shown):
+        column_names_to_print += f" (and {len(text_names) - len(shown)} more)"
+
+    warnings.warn(
+        f"These columns look like free text and are being ordinal-encoded as "
+        f"high-cardinality categoricals, which usually adds noise rather than "
+        f"signal: {column_names_to_print}.\n"
+        "If such a column holds numbers stored as strings, convert it to a numeric "
+        "dtype. If it holds genuine text, this package has no text handling -- "
+        "consider the tabpfn-client API, which embeds text natively: "
+        "https://github.com/PriorLabs/tabpfn-client \n"
+        "To silence this for a column that is genuinely a high-cardinality category, "
+        "pass its index in `categorical_features_indices`.",
+        UserWarning,
+        # Points at a direct `estimator.fit(X, y)` call site. Six frames out: this
+        # function, `detect_feature_modalities`, `_initialize_dataset_preprocessing`,
+        # `fit`, and the contextlib wrapper added by the `@config_context(...)`
+        # decorator on `fit`. Pinned by the `warning.filename` asserts in the tests.
+        stacklevel=6,
+    )
 
 
 def _detect_feature_modality(
